# apps/web/public/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job Seek API",
    "version": "1.4.0",
...
  },
  "servers": [{ "url": "https://jseek.co" }],
  "paths": {
    "/api/v1/feedback": {
      "post": {
        "operationId": "submitFeedback",
        "summary": "Submit product feedback, a bug report or feature suggestion",
        "description": "Optional anonymous feedback about an actual Job Seek task. Omit personal information, credentials, private URLs, raw tool arguments and transcripts. Submissions are stored with server version and timestamp; no tracking ID, deduplication or status/read API. Limit 5/hour per client IP; maximum body 16 KiB. Retrying can create repeated submissions.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "goal": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 500,
                        "description": "Task you were trying to accomplish, summarized without personal information or the original conversation"
                      },
                      "affectedTool": {
                        "description": "Job Seek tool involved, if applicable",
                        "type": "string",
                        "enum": [
                          "search_jobs",
                          "get_job_detail",
                          "search_companies",
                          "list_taxonomies",
                          "resolve_slugs",
                          "create_watchlist_link",
                          "report_bug",
                          "suggest_feature",
                          "give_feedback"
                        ]
                      },
                      "expected": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 1000,
                        "description": "Expected behavior"
                      },
                      "observed": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 1000,
                        "description": "What actually happened; distinguish observations from assumptions"
                      },
                      "reproduction": {
                        "description": "Minimal sanitized reproduction; no raw arguments, credentials or transcripts",
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 1000
                      },
                      "impact": {
                        "type": "string",
                        "enum": [
                          "blocked",
                          "partial",
                          "inconvenience"
                        ],
                        "description": "How the problem affected completion of the task"
                      },
                      "workaround": {
                        "description": "Workaround, if any",
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 500
                      },
                      "kind": {
                        "type": "string",
                        "const": "bug"
                      }
                    },
                    "required": [
                      "goal",
                      "expected",
                      "observed",
                      "impact",
                      "kind"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "goal": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 500,
                        "description": "Task you were trying to accomplish, summarized without personal information or the original conversation"
                      },
                      "affectedTool": {
                        "description": "Job Seek tool involved, if applicable",
                        "type": "string",
                        "enum": [
                          "search_jobs",
                          "get_job_detail",
                          "search_companies",
                          "list_taxonomies",
                          "resolve_slugs",
                          "create_watchlist_link",
                          "report_bug",
                          "suggest_feature",
                          "give_feedback"
                        ]
                      },
                      "capability": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 1000,
                        "description": "Concrete missing capability"
                      },
                      "benefit": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 1000,
                        "description": "How the capability would improve the task outcome"
                      },
                      "impact": {
                        "type": "string",
                        "enum": [
                          "blocked",
                          "partial",
                          "inconvenience"
                        ],
                        "description": "How the problem affected completion of the task"
                      },
                      "workaround": {
                        "description": "How you currently work around the limitation",
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 500
                      },
                      "kind": {
                        "type": "string",
                        "const": "feature"
                      }
                    },
                    "required": [
                      "goal",
                      "capability",
                      "benefit",
                      "impact",
                      "kind"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "goal": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 500,
                        "description": "Task you were trying to accomplish, summarized without personal information or the original conversation"
                      },
                      "affectedTool": {
                        "description": "Job Seek tool involved, if applicable",
                        "type": "string",
                        "enum": [
                          "search_jobs",
                          "get_job_detail",
                          "search_companies",
                          "list_taxonomies",
                          "resolve_slugs",
                          "create_watchlist_link",
                          "report_bug",
                          "suggest_feature",
                          "give_feedback"
                        ]
                      },
                      "observation": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 1000,
                        "description": "What was useful or confusing, based on actual use"
                      },
                      "improvement": {
                        "description": "Optional suggested improvement",
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 1000
                      },
                      "kind": {
                        "type": "string",
                        "const": "feedback"
                      }
                    },
                    "required": [
                      "goal",
                      "observation",
                      "kind"
                    ],
                    "additionalProperties": false
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Submission stored",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "const": true
                    }
                  },
                  "required": [
                    "success"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid submission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "const": false
                    },
                    "error": {
                      "type": "string",
                      "enum": [
                        "invalid_submission",
                        "too_large",
                        "unsupported_media_type",
                        "rate_limited",
                        "temporarily_unavailable"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "const": false
                    },
                    "error": {
                      "type": "string",
                      "enum": [
                        "invalid_submission",
                        "too_large",
                        "unsupported_media_type",
                        "rate_limited",
                        "temporarily_unavailable"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "415": {
            "description": "JSON content type required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "const": false
                    },
                    "error": {
                      "type": "string",
                      "enum": [
                        "invalid_submission",
                        "too_large",
                        "unsupported_media_type",
                        "rate_limited",
                        "temporarily_unavailable"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Feedback submission rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "const": false
                    },
                    "error": {
                      "type": "string",
                      "enum": [
                        "invalid_submission",
                        "too_large",
                        "unsupported_media_type",
                        "rate_limited",
                        "temporarily_unavailable"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds until a later attempt is allowed",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "503": {
            "description": "Submission temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "const": false
                    },
                    "error": {
                      "type": "string",
                      "enum": [
                        "invalid_submission",
                        "too_large",
                        "unsupported_media_type",
                        "rate_limited",
                        "temporarily_unavailable"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}